otomatisasi normalisasi kriteria

+++ database/migrations/2023_10_17_192305_create_selection_criterias_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->integer('weight');
 
             // $table->longText('description')->nullable();
-            $table->float('weight_normalization')->nullable();
+            $table->decimal('weight_normalization', 10,9)->nullable();
             
             $table->unsignedBigInteger('selection_id');
             $table->foreign('selection_id')->references('id')->on('selections')->onDelete("cascade");
